% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_agencies_no_admin}
\alias{get_agencies_no_admin}
\title{Return agencies that don't have an admin
Check if the agency admins are active}
\usage{
get_agencies_no_admin(
  agency = agency,
  agency_admin_address = agency_admin_address
)
}
\arguments{
\item{agency}{a dataframe}

\item{agency_admin_address}{a dataframe}
}
\value{
Returns a message and list of agencies
}
\description{
Return agencies that don't have an admin
Check if the agency admins are active
}
\examples{
get_agencies_no_admin <- function(agency = agency,
                                  agency_admin_address = agency_admin_address)
}
