% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_users_not_in_elearning}
\alias{get_users_not_in_elearning}
\title{Check if any active HMIS users are not in eLearning}
\usage{
get_users_not_in_elearning(users = NULL, user_course_report = NULL)
}
\arguments{
\item{users}{a dataframe}

\item{user_course_report}{a dataframe}
}
\value{
Returns error and list if any active users aren't in eLearning
}
\description{
Check if any active HMIS users are not in eLearning
}
\examples{
get_users_not_in_elearning(users = files$users,
                           user_course_report = files$user_course_report)
}
