% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_users_overdue_renewal}
\alias{get_users_overdue_renewal}
\title{Check HMIS users overdue for renewal; give results to Monica}
\usage{
get_users_overdue_renewal(users = NULL, user_course_report = NULL)
}
\arguments{
\item{users}{a dataframe}

\item{user_course_report}{a dataframe}
}
\value{
Returns message and list of users
}
\description{
Check HMIS users overdue for renewal; give results to Monica
}
\examples{
get_users_overdue_renewal(users = users,
                           user_course_report = user_course_report)
}
