% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_csv_files}
\alias{load_csv_files}
\title{Load Looker files and eLearning file}
\usage{
load_csv_files()
}
\value{
A list of dataframes.
}
\description{
Load Looker files and eLearning file
}
\examples{
files <- load_csv_files()
}
